Refactor and optimize page layouts and scripts
The code refactor focuses on page layouts and scripts across the project. In detail, it includes the simplification of the Index view in home, removal of unnecessary developer data-attribute in modal-body image, and js section restructuring. Furthermore, the shared Layout was significantly changed with a new page structure and meta tags, scripts, and links updated. The package.json file was also updated with new dependencies, and a new auto-theme script was added. Lastly, the AiurDriveDbContext.cs file was simplified.