Update file saving process and modify navbar
The main update is the enhancement of the file saving process in the StorageService. Now, files are saved according to the current date and time, and mechanisms have been added to check for existing files and prevent overwriting. Also, there has been a modification in the navbar, changing "Player Name" to "Sign in".