Add Docker to installation and update image scaling
The install script has been updated to include Docker in the prerequisites and to modify the user groups. The handling of hyper-scaled images in the download controller and the hyper-scale service was updated to improve efficiency and accuracy. The Docker requirements have also been added to the README.md file.